tests: move old CRD validation tests to test/crdsvalidation/
