+++ test/envtest/crds_envtest_test.go
@@ -2,22 +2,14 @@ package envtest
 
 import (
 	"context"
-	"fmt"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/samber/lo"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap/zaptest/observer"
-	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	configurationv1 "github.com/kong/kong-operator/v2/api/configuration/v1"
-	configurationv1alpha1 "github.com/kong/kong-operator/v2/api/configuration/v1alpha1"
-	configurationv1beta1 "github.com/kong/kong-operator/v2/api/configuration/v1beta1"
 	"github.com/kong/kong-operator/v2/ingress-controller/pkg/manager"
 	managercfg "github.com/kong/kong-operator/v2/ingress-controller/pkg/manager/config"
 )
@@ -103,799 +95,4 @@ func TestNoKongCRDsInstalledIsFatal(t *testing.T) {
 	require.NoError(t, err)
 
 	require.ErrorContains(t, m.Run(ctx), "timed out waiting for cache to be synced")
-}
-
-func TestCRDValidations(t *testing.T) {
-	t.Parallel()
-
-	ctx := t.Context()
-	scheme := Scheme(t, WithKong)
-	envcfg, _ := Setup(t, ctx, scheme)
-	ctrlClient := NewControllerClient(t, scheme, envcfg)
-
-	testCases := []struct {
-		name     string
-		scenario func(ctx context.Context, t *testing.T, ns string)
-	}{
-		{
-			name: "KongUpstreamPolicy - only one of spec.hashOn.(input|cookie|header|uriCapture|queryArg) can be set",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				for i, invalidHashOn := range generateInvalidHashOns() {
-					t.Run(fmt.Sprintf("invalidHashOn[%d]", i), func(t *testing.T) {
-						err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-							HashOn:    &invalidHashOn,
-							Algorithm: new("consistent-hashing"),
-						})
-						require.ErrorContains(t, err, "Only one of spec.hashOn.(input|cookie|header|uriCapture|queryArg) can be set.")
-					})
-				}
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - only one of spec.hashOnFallback.(input|header|uriCapture|queryArg) can be set",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				invalidHashOns := lo.Reject(generateInvalidHashOns(), func(hashOn configurationv1beta1.KongUpstreamHash, _ int) bool {
-					// Filter out Cookie which is not allowed in spec.hashOnFallback.
-					return hashOn.Cookie != nil
-				})
-				for i, invalidHashOn := range invalidHashOns {
-					t.Run(fmt.Sprintf("invalidHashOn[%d]", i), func(t *testing.T) {
-						err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-							HashOnFallback: &invalidHashOn,
-						})
-						require.ErrorContains(t, err, "Only one of spec.hashOnFallback.(input|header|uriCapture|queryArg) can be set.")
-					})
-				}
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - spec.hashOn.cookie and spec.hashOn.cookiePath are set",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Algorithm: new("consistent-hashing"),
-					HashOn: &configurationv1beta1.KongUpstreamHash{
-						Cookie:     new("cookie-name"),
-						CookiePath: new("/"),
-					},
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - spec.hashOn.cookie is set, spec.hashOn.cookiePath is required",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Algorithm: new("consistent-hashing"),
-					HashOn: &configurationv1beta1.KongUpstreamHash{
-						Cookie: new("cookie-name"),
-					},
-				})
-				require.ErrorContains(t, err, "When spec.hashOn.cookie is set, spec.hashOn.cookiePath is required.")
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - spec.hashOn.cookiePath is set, spec.hashOn.cookie is required",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Algorithm: new("consistent-hashing"),
-					HashOn: &configurationv1beta1.KongUpstreamHash{
-						CookiePath: new("/"),
-					},
-				})
-				require.ErrorContains(t, err, "When spec.hashOn.cookiePath is set, spec.hashOn.cookie is required.")
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - spec.hashOnFallback.cookie can be set when spec.hashOn uses header",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Algorithm: new("consistent-hashing"),
-					HashOn: &configurationv1beta1.KongUpstreamHash{
-						Header: new("header-name"),
-					},
-					HashOnFallback: &configurationv1beta1.KongUpstreamHash{
-						Cookie:     new("cookie-name"),
-						CookiePath: new("/"),
-					},
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - spec.hashOnFallback.cookie can be set when spec.hashOn uses queryArg",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Algorithm: new("consistent-hashing"),
-					HashOn: &configurationv1beta1.KongUpstreamHash{
-						QueryArg: new("query-arg-name"),
-					},
-					HashOnFallback: &configurationv1beta1.KongUpstreamHash{
-						Cookie:     new("cookie-name"),
-						CookiePath: new("/"),
-					},
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - spec.hashOnFallback.cookie can be set when spec.hashOn uses uriCapture",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Algorithm: new("consistent-hashing"),
-					HashOn: &configurationv1beta1.KongUpstreamHash{
-						URICapture: new("uri-capture-name"),
-					},
-					HashOnFallback: &configurationv1beta1.KongUpstreamHash{
-						Cookie:     new("cookie-name"),
-						CookiePath: new("/"),
-					},
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			// Ref: https://developer.konghq.com/gateway/entities/upstream/#consistent-hashing
-			// > The hash_fallback setting is invalid and can’t be used if cookie is the primary hashing mechanism.
-			name: "KongUpstreamPolicy - spec.hashOnFallback.cookie can't be set when spec.hashOn uses cookie",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Algorithm: new("consistent-hashing"),
-					HashOn: &configurationv1beta1.KongUpstreamHash{
-						Cookie:     new("cookie-name"),
-						CookiePath: new("/cookie"),
-					},
-					HashOnFallback: &configurationv1beta1.KongUpstreamHash{
-						Cookie:     new("cookie-name-2"),
-						CookiePath: new("/cookie-2"),
-					},
-				})
-				require.ErrorContains(t, err, "spec.hashOnFallback must not be set when spec.hashOn.cookie is set.")
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - spec.hashOn.uri_capture can be used with spec.hashOnFallback.header",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Algorithm: new("consistent-hashing"),
-					HashOn: &configurationv1beta1.KongUpstreamHash{
-						URICapture: new("uri-capture-name"),
-					},
-					HashOnFallback: &configurationv1beta1.KongUpstreamHash{
-						Header: new("header-name"),
-					},
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - healthchecks.active.healthy.httpStatuses contains invalid HTTP status code",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Healthchecks: &configurationv1beta1.KongUpstreamHealthcheck{
-						Active: &configurationv1beta1.KongUpstreamActiveHealthcheck{
-							Healthy: &configurationv1beta1.KongUpstreamHealthcheckHealthy{
-								HTTPStatuses: []configurationv1beta1.HTTPStatus{600},
-							},
-						},
-					},
-				})
-				require.ErrorContains(t, err, "should be less than or equal to 599")
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - healthchecks.active.unhealthy.httpStatuses contains invalid HTTP status code",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Healthchecks: &configurationv1beta1.KongUpstreamHealthcheck{
-						Active: &configurationv1beta1.KongUpstreamActiveHealthcheck{
-							Unhealthy: &configurationv1beta1.KongUpstreamHealthcheckUnhealthy{
-								HTTPStatuses: []configurationv1beta1.HTTPStatus{99},
-							},
-						},
-					},
-				})
-				require.ErrorContains(t, err, "should be greater than or equal to 100")
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - healthchecks.passive.healthy.interval must not be set",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Healthchecks: &configurationv1beta1.KongUpstreamHealthcheck{
-						Passive: &configurationv1beta1.KongUpstreamPassiveHealthcheck{
-							Healthy: &configurationv1beta1.KongUpstreamHealthcheckHealthy{
-								Interval: new(10),
-							},
-						},
-					},
-				})
-				require.ErrorContains(t, err, "spec.healthchecks.passive.healthy.interval must not be set.")
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - healthchecks.passive.unhealthy.interval must not be set",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Healthchecks: &configurationv1beta1.KongUpstreamHealthcheck{
-						Passive: &configurationv1beta1.KongUpstreamPassiveHealthcheck{
-							Unhealthy: &configurationv1beta1.KongUpstreamHealthcheckUnhealthy{
-								Interval: new(10),
-							},
-						},
-					},
-				})
-				require.ErrorContains(t, err, "spec.healthchecks.passive.unhealthy.interval must not be set.")
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - hashOn can only be set when algorithm is set to consistent-hashing or sticky-sessions",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					HashOn: &configurationv1beta1.KongUpstreamHash{
-						Header: new("header-name"), // Could be any of the hashOn fields.
-					},
-				})
-				require.ErrorContains(t, err, `spec.algorithm must be set to either 'consistent-hashing' or 'sticky-sessions' when spec.hashOn is set.`)
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - hashOnFallback can only be set when algorithm is set to consistent-hashing",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					HashOnFallback: &configurationv1beta1.KongUpstreamHash{
-						Header: new("header-name"), // Could be any of the hashOn fields.
-					},
-				})
-				require.ErrorContains(t, err, `spec.algorithm must be set to "consistent-hashing" when spec.hashOnFallback is set.`)
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - hashOn(Fallback).input enum is validated",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				validValues := []string{"ip", "consumer", "path", "none"}
-				for _, validValue := range validValues {
-					t.Run(fmt.Sprintf("valid-value[%s]", validValue), func(t *testing.T) {
-						err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-							Algorithm: new("consistent-hashing"),
-							HashOn: &configurationv1beta1.KongUpstreamHash{
-								Input: new(configurationv1beta1.HashInput(validValue)),
-							},
-							HashOnFallback: &configurationv1beta1.KongUpstreamHash{
-								Input: new(configurationv1beta1.HashInput(validValue)),
-							},
-						})
-						require.NoError(t, err)
-					})
-				}
-
-				t.Run("invalid value", func(t *testing.T) {
-					err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-						Algorithm: new("consistent-hashing"),
-						HashOn: &configurationv1beta1.KongUpstreamHash{
-							Input: new(configurationv1beta1.HashInput("unknown-input")),
-						},
-						HashOnFallback: &configurationv1beta1.KongUpstreamHash{
-							Input: new(configurationv1beta1.HashInput("unknown-input-fallback")),
-						},
-					})
-					require.ErrorContains(t, err, `spec.hashOn.input: Unsupported value: "unknown-input": supported values: "ip", "consumer", "path"`)
-					require.ErrorContains(t, err, `spec.hashOnFallback.input: Unsupported value: "unknown-input-fallback": supported values: "ip", "consumer", "path"`)
-				})
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - algorithm enum is validated",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				validValues := []string{"consistent-hashing", "round-robin", "least-connections", "latency", "sticky-sessions"}
-				for _, validValue := range validValues {
-					t.Run(fmt.Sprintf("valid-value[%s]", validValue), func(t *testing.T) {
-						err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-							Algorithm: new(validValue),
-						})
-						require.NoError(t, err)
-					})
-				}
-
-				t.Run("invalid value", func(t *testing.T) {
-					err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-						Algorithm: new("unknown-algorithm"),
-					})
-					require.ErrorContains(t, err, `spec.algorithm: Unsupported value: "unknown-algorithm": supported values: "round-robin", "consistent-hashing", "least-connections", "latency", "sticky-sessions"`)
-				})
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - healthcheck.(active|passive).type enum is validated",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				validValues := []string{"http", "https", "tcp", "grpc", "grpcs"}
-				for _, validValue := range validValues {
-					t.Run(fmt.Sprintf("valid-value[%s]", validValue), func(t *testing.T) {
-						err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-							Healthchecks: &configurationv1beta1.KongUpstreamHealthcheck{
-								Active: &configurationv1beta1.KongUpstreamActiveHealthcheck{
-									Type: new(validValue),
-								},
-								Passive: &configurationv1beta1.KongUpstreamPassiveHealthcheck{
-									Type: new(validValue),
-								},
-							},
-						})
-						require.NoError(t, err)
-					})
-				}
-				t.Run("invalid value", func(t *testing.T) {
-					err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-						Healthchecks: &configurationv1beta1.KongUpstreamHealthcheck{
-							Active: &configurationv1beta1.KongUpstreamActiveHealthcheck{
-								Type: new("unknown-type-active"),
-							},
-							Passive: &configurationv1beta1.KongUpstreamPassiveHealthcheck{
-								Type: new("unknown-type-passive"),
-							},
-						},
-					})
-					require.ErrorContains(t, err, `spec.healthchecks.active.type: Unsupported value: "unknown-type-active": supported values: "http", "https", "tcp", "grpc", "grpcs"`)
-					require.ErrorContains(t, err, `spec.healthchecks.passive.type: Unsupported value: "unknown-type-passive": supported values: "http", "https", "tcp", "grpc", "grpcs"`)
-				})
-			},
-		},
-		{
-			name: "KongUpstreamPolicy - hashOnFallback must not be set when spec.hasOn.cookie is set",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongUpstreamPolicy(ctx, ctrlClient, ns, configurationv1beta1.KongUpstreamPolicySpec{
-					Algorithm: new("consistent-hashing"),
-					HashOn: &configurationv1beta1.KongUpstreamHash{
-						Cookie:     new("cookie-name"),
-						CookiePath: new("/"),
-					},
-					HashOnFallback: &configurationv1beta1.KongUpstreamHash{
-						Header: new("header-name"),
-					},
-				})
-				require.ErrorContains(t, err, `spec.hashOnFallback must not be set when spec.hashOn.cookie is set.`)
-			},
-		},
-		{
-			name: "KongPlugin - no config is allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongPlugin(ctx, ctrlClient, ns, &configurationv1.KongPlugin{
-					PluginName: "key-auth",
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "KongPlugin - with config is allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongPlugin(ctx, ctrlClient, ns, &configurationv1.KongPlugin{
-					PluginName: "key-auth",
-					Config: apiextensionsv1.JSON{
-						Raw: []byte(`{"key_names":["apikey"]}`),
-					},
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "KongPlugin - with configFrom is allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongPlugin(ctx, ctrlClient, ns, &configurationv1.KongPlugin{
-					PluginName: "key-auth",
-					ConfigFrom: &configurationv1.ConfigSource{
-						SecretValue: configurationv1.SecretValueFromSource{
-							Secret: "secret-name",
-							Key:    "key-name",
-						},
-					},
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "KongPlugin - with configFrom and config is rejected",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongPlugin(ctx, ctrlClient, ns, &configurationv1.KongPlugin{
-					PluginName: "key-auth",
-					Config: apiextensionsv1.JSON{
-						Raw: []byte(`{"key_names":["apikey"]}`),
-					},
-					ConfigFrom: &configurationv1.ConfigSource{
-						SecretValue: configurationv1.SecretValueFromSource{
-							Secret: "secret-name",
-							Key:    "key-name",
-						},
-					},
-				})
-				assert.Error(t, err)
-				assert.ErrorContains(t, err, "Using both config and configFrom fields is not allowed.")
-			},
-		},
-		{
-			name: "KongPlugin - change plugin field is not allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				plugin := &configurationv1.KongPlugin{
-					PluginName: "key-auth",
-				}
-				err := createKongPlugin(ctx, ctrlClient, ns, plugin)
-				require.NoError(t, err)
-				plugin.PluginName = "cors"
-				err = updateKongPlugin(ctx, ctrlClient, ns, plugin)
-				assert.Error(t, err)
-				assert.ErrorContains(t, err, "The plugin field is immutable")
-			},
-		},
-		{
-			name: "KongPlugin - using configPatches is allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				plugin := &configurationv1.KongPlugin{
-					PluginName: "key-auth",
-					ConfigPatches: []configurationv1.ConfigPatch{
-						{
-							Path: "/key_names",
-							ValueFrom: configurationv1.ConfigSource{
-								SecretValue: configurationv1.SecretValueFromSource{
-									Secret: "secret-name",
-									Key:    "key-name",
-								},
-							},
-						},
-					},
-				}
-				require.NoError(t, createKongPlugin(ctx, ctrlClient, ns, plugin))
-			},
-		},
-		{
-			name: "KongPlugin - using config and configPatches is allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				plugin := &configurationv1.KongPlugin{
-					PluginName: "key-auth",
-					Config: apiextensionsv1.JSON{
-						Raw: []byte(`{"key_name":"apikey"}`),
-					},
-					ConfigPatches: []configurationv1.ConfigPatch{
-						{
-							Path: "/key_names",
-							ValueFrom: configurationv1.ConfigSource{
-								SecretValue: configurationv1.SecretValueFromSource{
-									Secret: "secret-name",
-									Key:    "key-name",
-								},
-							},
-						},
-					},
-				}
-				require.NoError(t, createKongPlugin(ctx, ctrlClient, ns, plugin))
-			},
-		},
-		{
-			name: "KongPlugin - using configFrom and configPatches is not allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				plugin := &configurationv1.KongPlugin{
-					PluginName: "key-auth",
-					ConfigFrom: &configurationv1.ConfigSource{
-						SecretValue: configurationv1.SecretValueFromSource{
-							Secret: "secret-name",
-							Key:    "key-name",
-						},
-					},
-					ConfigPatches: []configurationv1.ConfigPatch{
-						{
-							Path: "/key_names",
-							ValueFrom: configurationv1.ConfigSource{
-								SecretValue: configurationv1.SecretValueFromSource{
-									Secret: "secret-name",
-									Key:    "key-name",
-								},
-							},
-						},
-					},
-				}
-				assert.ErrorContains(t, createKongPlugin(ctx, ctrlClient, ns, plugin),
-					"Using both configFrom and configPatches fields is not allowed.")
-			},
-		},
-		{
-			name: "KongClusterPlugin - no config is allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongClusterPlugin(ctx, ctrlClient, ns, &configurationv1.KongClusterPlugin{
-					PluginName: "key-auth",
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "KongClusterPlugin - with config is allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongClusterPlugin(ctx, ctrlClient, ns, &configurationv1.KongClusterPlugin{
-					PluginName: "key-auth",
-					Config: apiextensionsv1.JSON{
-						Raw: []byte(`{"key_names":["apikey"]}`),
-					},
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "KongClusterPlugin - with configFrom is allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongClusterPlugin(ctx, ctrlClient, ns, &configurationv1.KongClusterPlugin{
-					PluginName: "key-auth",
-					ConfigFrom: &configurationv1.NamespacedConfigSource{
-						SecretValue: configurationv1.NamespacedSecretValueFromSource{
-							Secret:    "secret-name",
-							Key:       "key-name",
-							Namespace: "ns",
-						},
-					},
-				})
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "KongClusterPlugin - with configFrom and config is rejected",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				err := createKongClusterPlugin(ctx, ctrlClient, ns, &configurationv1.KongClusterPlugin{
-					PluginName: "key-auth",
-					Config: apiextensionsv1.JSON{
-						Raw: []byte(`{"key_names":["apikey"]}`),
-					},
-					ConfigFrom: &configurationv1.NamespacedConfigSource{
-						SecretValue: configurationv1.NamespacedSecretValueFromSource{
-							Secret:    "secret-name",
-							Key:       "key-name",
-							Namespace: "ns",
-						},
-					},
-				})
-				assert.Error(t, err)
-				assert.ErrorContains(t, err, "Using both config and configFrom fields is not allowed.")
-			},
-		},
-		{
-			name: "KongClusterPlugin - change plugin field is not allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				plugin := &configurationv1.KongClusterPlugin{
-					PluginName: "key-auth",
-				}
-				err := createKongClusterPlugin(ctx, ctrlClient, ns, plugin)
-				require.NoError(t, err)
-				plugin.PluginName = "cors"
-				err = updateKongClusterPlugin(ctx, ctrlClient, ns, plugin)
-				assert.Error(t, err)
-				assert.ErrorContains(t, err, "The plugin field is immutable")
-			},
-		},
-		{
-			name: "KongClusterPlugin - using configPatches is allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				plugin := &configurationv1.KongClusterPlugin{
-					PluginName: "key-auth",
-					ConfigPatches: []configurationv1.NamespacedConfigPatch{
-						{
-							Path: "/key_names",
-							ValueFrom: configurationv1.NamespacedConfigSource{
-								SecretValue: configurationv1.NamespacedSecretValueFromSource{
-									Namespace: ns,
-									Secret:    "secret-name",
-									Key:       "key-name",
-								},
-							},
-						},
-					},
-				}
-				require.NoError(t, createKongClusterPlugin(ctx, ctrlClient, ns, plugin))
-			},
-		},
-		{
-			name: "KongClusterPlugin - using config and configPatches is allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				plugin := &configurationv1.KongClusterPlugin{
-					PluginName: "key-auth",
-					Config: apiextensionsv1.JSON{
-						Raw: []byte(`{"key_name":"apikey"}`),
-					},
-					ConfigPatches: []configurationv1.NamespacedConfigPatch{
-						{
-							Path: "/key_names",
-							ValueFrom: configurationv1.NamespacedConfigSource{
-								SecretValue: configurationv1.NamespacedSecretValueFromSource{
-									Namespace: ns,
-									Secret:    "secret-name",
-									Key:       "key-name",
-								},
-							},
-						},
-					},
-				}
-				require.NoError(t, createKongClusterPlugin(ctx, ctrlClient, ns, plugin))
-			},
-		},
-		{
-			name: "KongClusterPlugin - using configFrom and configPatches is not allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				plugin := &configurationv1.KongClusterPlugin{
-					PluginName: "key-auth",
-					ConfigFrom: &configurationv1.NamespacedConfigSource{
-						SecretValue: configurationv1.NamespacedSecretValueFromSource{
-							Secret: "secret-name",
-							Key:    "key-name",
-						},
-					},
-					ConfigPatches: []configurationv1.NamespacedConfigPatch{
-						{
-							Path: "/key_names",
-							ValueFrom: configurationv1.NamespacedConfigSource{
-								SecretValue: configurationv1.NamespacedSecretValueFromSource{
-									Namespace: ns,
-									Secret:    "secret-name",
-									Key:       "key-name",
-								},
-							},
-						},
-					},
-				}
-				assert.ErrorContains(t, createKongClusterPlugin(ctx, ctrlClient, ns, plugin),
-					"Using both configFrom and configPatches fields is not allowed.")
-			},
-		},
-		{
-			name: "KongVault - changing spec.prefix is not allowed",
-			scenario: func(ctx context.Context, t *testing.T, _ string) {
-				vault := &configurationv1alpha1.KongVault{
-					Spec: configurationv1alpha1.KongVaultSpec{
-						Backend: "env",
-						Prefix:  "env-0",
-					},
-				}
-				require.NoError(t, createKongVault(ctx, ctrlClient, vault))
-				vault.Spec.Prefix = "env-1"
-				assert.ErrorContains(t, updateKongVault(ctx, ctrlClient, vault),
-					"The spec.prefix field is immutable")
-			},
-		},
-		{
-			name: "KongConsumer - duplicate credentials are not allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				consumer := &configurationv1.KongConsumer{
-					Username:    "u1",
-					Credentials: []string{"c1", "c1", "c2"},
-				}
-				assert.ErrorContains(t, createKongConsumer(ctx, ctrlClient, ns, consumer), `Duplicate value: "c1"`)
-			},
-		},
-		{
-			name: "KongConsumer - unique credentials are allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				consumer := &configurationv1.KongConsumer{
-					Username:    "u1",
-					Credentials: []string{"c1", "c2"},
-				}
-				assert.NoError(t, createKongConsumer(ctx, ctrlClient, ns, consumer))
-			},
-		},
-		{
-			name: "KongConsumer - duplicate consumer groups are not allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				consumer := &configurationv1.KongConsumer{
-					Username:       "u1",
-					ConsumerGroups: []string{"cg1", "cg1", "cg2"},
-				}
-				assert.ErrorContains(t, createKongConsumer(ctx, ctrlClient, ns, consumer), `Duplicate value: "cg1"`)
-			},
-		},
-		{
-			name: "KongConsumer - unique consumer groups are allowed",
-			scenario: func(ctx context.Context, t *testing.T, ns string) {
-				consumer := &configurationv1.KongConsumer{
-					Username:       "u1",
-					ConsumerGroups: []string{"cg1", "cg2"},
-				}
-				assert.NoError(t, createKongConsumer(ctx, ctrlClient, ns, consumer))
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			ns := CreateNamespace(ctx, t, ctrlClient)
-			tc.scenario(ctx, t, ns.Name)
-		})
-	}
-}
-
-func createKongUpstreamPolicy(ctx context.Context, client client.Client, ns string, spec configurationv1beta1.KongUpstreamPolicySpec) error {
-	return client.Create(ctx, &configurationv1beta1.KongUpstreamPolicy{
-		ObjectMeta: metav1.ObjectMeta{
-			GenerateName: "test-",
-			Namespace:    ns,
-		},
-		Spec: spec,
-	})
-}
-
-// generateInvalidHashOns generates a list of KongUpstreamHash objects with all possible invalid fields pairs.
-func generateInvalidHashOns() []configurationv1beta1.KongUpstreamHash {
-	fieldSetFns := []func(h *configurationv1beta1.KongUpstreamHash){
-		func(h *configurationv1beta1.KongUpstreamHash) {
-			h.Input = new(configurationv1beta1.HashInput("consumer"))
-		},
-		func(h *configurationv1beta1.KongUpstreamHash) {
-			h.Cookie = new("cookie-name")
-			h.CookiePath = new("/")
-		},
-		func(h *configurationv1beta1.KongUpstreamHash) {
-			h.Header = new("header-name")
-		},
-		func(h *configurationv1beta1.KongUpstreamHash) {
-			h.URICapture = new("uri-capture")
-		},
-		func(h *configurationv1beta1.KongUpstreamHash) {
-			h.QueryArg = new("query-arg")
-		},
-	}
-
-	var invalidHashOns []configurationv1beta1.KongUpstreamHash
-	for outerIdx, fieldSetFn := range fieldSetFns {
-		hashOn := configurationv1beta1.KongUpstreamHash{}
-		fieldSetFn(&hashOn)
-
-		for innerIdx, innerFieldSetFn := range fieldSetFns {
-			if outerIdx == innerIdx {
-				continue
-			}
-			invalidHashOn := hashOn.DeepCopy()
-			innerFieldSetFn(invalidHashOn)
-			invalidHashOns = append(invalidHashOns, *invalidHashOn)
-		}
-	}
-
-	optStr := func(s *string) string {
-		if s == nil {
-			return "<nil>"
-		}
-		return *s
-	}
-	return lo.UniqBy(invalidHashOns, func(h configurationv1beta1.KongUpstreamHash) string {
-		return fmt.Sprintf("%s.%s.%s.%s", optStr(h.Cookie), optStr(h.Header), optStr(h.URICapture), optStr(h.QueryArg))
-	})
-}
-
-func createKongPlugin(ctx context.Context, client client.Client, ns string, plugin *configurationv1.KongPlugin) error {
-	plugin.GenerateName = "test-"
-	plugin.Namespace = ns
-	return client.Create(ctx, plugin)
-}
-
-func createKongClusterPlugin(ctx context.Context, client client.Client, ns string, plugin *configurationv1.KongClusterPlugin) error {
-	plugin.GenerateName = "test-"
-	plugin.Namespace = ns
-	return client.Create(ctx, plugin)
-}
-
-func updateKongPlugin(ctx context.Context, client client.Client, ns string, plugin *configurationv1.KongPlugin) error {
-	plugin.GenerateName = "test-"
-	plugin.Namespace = ns
-	return client.Update(ctx, plugin)
-}
-
-func updateKongClusterPlugin(ctx context.Context, client client.Client, ns string, plugin *configurationv1.KongClusterPlugin) error {
-	plugin.GenerateName = "test-"
-	plugin.Namespace = ns
-	return client.Update(ctx, plugin)
-}
-
-func createKongVault(ctx context.Context, c client.Client, vault *configurationv1alpha1.KongVault) error {
-	vault.GenerateName = "test-"
-	return c.Create(ctx, vault)
-}
-
-func updateKongVault(ctx context.Context, c client.Client, vault *configurationv1alpha1.KongVault) error {
-	vault.GenerateName = "test-"
-	return c.Update(ctx, vault)
-}
-
-func createKongConsumer(ctx context.Context, c client.Client, ns string, consumer *configurationv1.KongConsumer) error {
-	consumer.GenerateName = "test-"
-	consumer.Namespace = ns
-	return c.Create(ctx, consumer)
 }
